requirements.txt
streamlit
pandas
requests
schedule
python-dotenv
plotly